% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFunctions.R
\name{animateCollision}
\alias{animateCollision}
\title{Plot an animation for the collisions between balls.}
\usage{
animateCollision(posDF, step = Inf)
}
\arguments{
\item{posDF}{Data.frame with ball positions.}

\item{step}{Numeric with step.
step = Inf preserves 'usual' animation, otherwise 'shiny' animation is used}
}
\value{
No return value, a plot is drawn to the current device.
}
\description{
Plot an animation for the collisions between balls.
}
