% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFunctions.R
\name{animateThrow}
\alias{animateThrow}
\title{Plot the animation for a specific ball throw.}
\usage{
animateThrow(distance, color, step = Inf)
}
\arguments{
\item{distance}{Numeric with distance at which the ball is thrown.}

\item{color}{String with color.}

\item{step}{Numeric with step.
step = Inf preserves 'usual' animation, otherwise 'shiny' animation is used}
}
\value{
No return value, a plot is drawn to the current device.
}
\description{
Plot the animation for a specific ball throw.
}
