% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{computePWe}
\alias{computePWe}
\title{Special rating algorithm for unrated players, 
players with less than 8 games and player with all wins or all losses}
\usage{
computePWe(rankingPlayer, rankingPlayeri)
}
\arguments{
\item{rankingPlayer}{Numeric with ranking player 1.}

\item{rankingPlayeri}{Numeric with ranking player 2.}
}
\description{
Special rating algorithm for unrated players, 
players with less than 8 games and player with all wins or all losses
}
