% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineOutcome.R
\name{determineOutcome}
\alias{determineOutcome}
\title{Determine the outcome of the game.}
\usage{
determineOutcome(posDF)
}
\arguments{
\item{posDF}{Data.frame with ball positions.}
}
\value{
List with: 'winner' and 'pointsWon'
}
\description{
Determine the outcome of the game.
}
