% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawBernoulli}
\alias{drawBernoulli}
\title{Draw Bernouilli distribution.}
\usage{
drawBernoulli(param1)
}
\arguments{
\item{param1}{Numeric with first parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw Bernouilli distribution.
}
