% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawBinomial}
\alias{drawBinomial}
\title{Draw binomial distribution.}
\usage{
drawBinomial(param1, param2)
}
\arguments{
\item{param1}{Numeric with first parameter from the distribution.}

\item{param2}{Numeric with second parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw binomial distribution.
}
