% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawChiSquared}
\alias{drawChiSquared}
\title{Draw Chi-squared distribution.}
\usage{
drawChiSquared(param1)
}
\arguments{
\item{param1}{Numeric with first parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw Chi-squared distribution.
}
