% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawDistribution}
\alias{drawDistribution}
\title{Draw statistical distributions.}
\usage{
drawDistribution(distribution, param1, param2)
}
\arguments{
\item{distribution}{string with distribution name, among:
'normal', 'uniform', 'poisson', 'bernoulli', 'binomial',
'geometric', 'degenerate', 'chisq', 'weibull'.}

\item{param1}{Numeric with first parameter from the distribution.}

\item{param2}{Numeric with second parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw statistical distributions.
}
