% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFunctions.R
\name{drawField}
\alias{drawField}
\title{Draw the field.}
\usage{
drawField(newPlot = FALSE)
}
\arguments{
\item{newPlot}{Logical, if TRUE (FALSE by default),
should a blank plot be created?}
}
\value{
No return value, a plot is drawn to the current device.
}
\description{
Draw the field.
}
