% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFunctions.R
\name{drawFlower}
\alias{drawFlower}
\title{Draw a flower.}
\usage{
drawFlower(x, stemHeight, colNum = 2)
}
\arguments{
\item{x}{Numeric with flower position}

\item{stemHeight}{Numeric with stem height.}

\item{colNum}{Numeric, number of the flower.}
}
\value{
No return value, a plot is drawn to the current device.
}
\description{
Draw a flower.
}
