% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawGaussian}
\alias{drawGaussian}
\title{Draw Gaussian distribution.}
\usage{
drawGaussian(param1 = 5, param2 = 2)
}
\arguments{
\item{param1}{Numeric with first parameter from the distribution.}

\item{param2}{Numeric with second parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw Gaussian distribution.
}
