% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawGeometric}
\alias{drawGeometric}
\title{Draw geometric distribution.}
\usage{
drawGeometric(param1)
}
\arguments{
\item{param1}{Numeric with first parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw geometric distribution.
}
