% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFunctions.R
\name{drawHuman}
\alias{drawHuman}
\title{Draw a human.}
\usage{
drawHuman(color = "orange")
}
\arguments{
\item{color}{String with color.}
}
\value{
No return value, a plot is drawn to the current device.
}
\description{
Draw a human.
}
