% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawPoisson}
\alias{drawPoisson}
\title{Draw Poisson distribution.}
\usage{
drawPoisson(param1)
}
\arguments{
\item{param1}{Numeric with first parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw Poisson distribution.
}
