% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawUniform}
\alias{drawUniform}
\title{Draw Uniform distribution.}
\usage{
drawUniform(param1 = 3, param2 = 7)
}
\arguments{
\item{param1}{Numeric with first parameter from the distribution.}

\item{param2}{Numeric with second parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw Uniform distribution.
}
