% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{drawWeibull}
\alias{drawWeibull}
\title{Draw Weibull distribution.}
\usage{
drawWeibull(param1, param2)
}
\arguments{
\item{param1}{Numeric with first parameter from the distribution.}

\item{param2}{Numeric with second parameter from the distribution.}
}
\value{
no returned value, a plot is drawn in the current device.
}
\description{
Draw Weibull distribution.
}
