% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{generateNegativeMessage}
\alias{generateNegativeMessage}
\title{Generate a negative message for the application.}
\usage{
generateNegativeMessage()
}
\value{
String with message.
}
\description{
Generate a negative message for the application.
}
