% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOptions.R
\name{generateOptions}
\alias{generateOptions}
\title{Generate a set of statistical distributions and associated parameters.}
\usage{
generateOptions(posDF = NULL, n = 3)
}
\arguments{
\item{posDF}{Data.frame with ball positions.}

\item{n}{Number of distributions/parameter sets to generate.}
}
\value{
Sub-list of length \code{n} 
with 'type' of distribution, 'param1': first parameter,
'param2': second parameter.
}
\description{
Generate a set of statistical distributions and associated parameters.
}
