% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{generatePositiveMessage}
\alias{generatePositiveMessage}
\title{Generate a 'positive' message for the application.}
\usage{
generatePositiveMessage(distance)
}
\arguments{
\item{distance}{Distance of the throw.}
}
\value{
String with message.
}
\description{
Generate a 'positive' message for the application.
}
