% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisionPetanque.R
\name{getDistanceFromSpeed}
\alias{getDistanceFromSpeed}
\title{Get distance from initial speed}
\usage{
getDistanceFromSpeed(v0, a = 10)
}
\arguments{
\item{v0}{Initial speed, in m/s}

\item{a}{Acceleration (currently fixed), decceleration of 10m/s^2 by default.}
}
\value{
Achieved distance, in m
}
\description{
Get distance from initial speed
}
\author{
Laure Cougnaud, Maxim Nazarov
}
