% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisionPetanque.R
\name{getMass}
\alias{getMass}
\title{Get Mass of a ball}
\usage{
getMass(type)
}
\arguments{
\item{type}{String with ball type}
}
\value{
Mass of the target in grams
}
\description{
Get Mass of a ball
}
\author{
Laure Cougnaud
}
