% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{oaTitlePanel}
\alias{oaTitlePanel}
\title{Create title panel.}
\usage{
oaTitlePanel(appName, pkgName, logo, logo2 = NULL, subtitle = NULL,
  version = TRUE)
}
\arguments{
\item{appName}{String with application name.}

\item{pkgName}{String with package name.}

\item{logo}{String, path to first logo.}

\item{logo2}{String, path to second logo.}

\item{subtitle}{Subtitle string}

\item{version}{Logical, if TRUE (by default) the application
version is included in the app.}
}
\value{
\code{\link[shiny]{titlePanel}}.
}
\description{
Create title panel.
}
