% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-server.R
\name{petanqueServer}
\alias{petanqueServer}
\title{Shiny server script.}
\usage{
petanqueServer(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
No return value
}
\description{
Shiny server script.
}
