% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{petanqueUI}
\alias{petanqueUI}
\title{Shiny UI script.}
\usage{
petanqueUI(debug = FALSE)
}
\arguments{
\item{debug}{Logical, should the debug functionality
be included in the app (FALSE by default).}
}
\value{
\code{\link[shiny]{fluidPage}}.
}
\description{
Shiny UI script.
}
