% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOptions.R
\name{pickOption}
\alias{pickOption}
\title{Pick a distribution from a set of statistical distributions/parameters,
and throw the ball with \code{\link{throwBall}}.}
\usage{
pickOption(options, optionSelected = 1, posDF)
}
\arguments{
\item{options}{List with set of statistical distributions/parameters,
as generated by: \code{\link{generateOptions}}.}

\item{optionSelected}{Integer with selected option,
1 by default.}

\item{posDF}{Data.frame with ball positions.}
}
\value{
Updated \code{posDF} after the throw.
}
\description{
Pick a distribution from a set of statistical distributions/parameters,
and throw the ball with \code{\link{throwBall}}.
}
