% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{playerInfoModal}
\alias{playerInfoModal}
\title{Print information on players in a modal dialog.}
\usage{
playerInfoModal(failed1 = FALSE, failed2 = FALSE, failedBoth = FALSE)
}
\arguments{
\item{failed1}{Logical, if TRUE (FALSE by default), player 1 has failed the game.}

\item{failed2}{Logical, if TRUE (FALSE by default), player 2 has failed the game.}

\item{failedBoth}{Logical, if TRUE (FALSE by default), both players have failed the game.}
}
\value{
\code{\link[shiny]{modalDialog}}
}
\description{
Print information on players in a modal dialog.
}
