% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOptions.R
\name{printDistr}
\alias{printDistr}
\title{Print information for a specific distribution.}
\usage{
printDistr(distr, inButton = FALSE)
}
\arguments{
\item{distr}{List with distribution 'type', parameters: 'param1' and 'param2'.}

\item{inButton}{Logical, if TRUE (FALSE by default)
should the info fit into a button?}
}
\value{
string with HTML code
}
\description{
Print information for a specific distribution.
}
