% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOptions.R
\name{printOptions}
\alias{printOptions}
\title{Print different options of statistical distributions/set of parameters.}
\usage{
printOptions(options)
}
\arguments{
\item{options}{List with set of statistical distributions/parameters,
as generated by: \code{\link{generateOptions}}.}
}
\value{
No return value, options are printed in the console.
}
\description{
Print different options of statistical distributions/set of parameters.
}
