% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{printRankings}
\alias{printRankings}
\title{Extract ranking info from the file.}
\usage{
printRankings(file = RANKING_FILE)
}
\arguments{
\item{file}{Path to the RDS file with saved ranking}
}
\value{
Data.frame with ranking info.
}
\description{
Extract ranking info from the file.
}
