% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{rankingDocumentation}
\alias{rankingDocumentation}
\title{Create documentation about the ranking in the application.}
\usage{
rankingDocumentation()
}
\value{
\code{\link[shiny]{modalDialog}}
}
\description{
Create documentation about the ranking in the application.
}
