% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFunctions.R
\name{redrawBalls}
\alias{redrawBalls}
\title{Redraw the balls based on the updated positions.}
\usage{
redrawBalls(posDF, except = NULL)
}
\arguments{
\item{posDF}{Data.frame with ball positions.}

\item{except}{Numeric with balls ID to not re-redraw.}
}
\value{
No return value, a plot is drawn to the current device.
}
\description{
Redraw the balls based on the updated positions.
}
