% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawFunctions.R
\name{refreshPlot}
\alias{refreshPlot}
\title{'Refresh' the plot for a petanque game.}
\usage{
refreshPlot(posDF, newPlot = FALSE)
}
\arguments{
\item{posDF}{Data.frame with ball positions.}

\item{newPlot}{Logical, if TRUE (FALSE by default),
should a blank plot be created?}
}
\value{
No return value, a plot is drawn to the current device.
}
\description{
'Refresh' the plot for a petanque game.
}
