% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{resetRanking}
\alias{resetRanking}
\title{Reset ranking.}
\usage{
resetRanking(file = RANKING_FILE)
}
\arguments{
\item{file}{Path to the RDS file with saved ranking}
}
\value{
No return value, the \code{file} is saved and removed..
}
\description{
Reset ranking.
}
