% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Launch 'Petanque' Shiny Application}
\usage{
runShinyApp(debug = FALSE, ...)
}
\arguments{
\item{debug}{whether to allow to connect to console for debugging}

\item{...}{arguments (eventually) passed to \code{\link[shiny]{runApp}}}
}
\description{
Launch 'Petanque' Shiny Application
}
\author{
Maxim Nazarov
}
