% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-server.R
\name{svgDevice}
\alias{svgDevice}
\title{Open default svg device.}
\usage{
svgDevice(...)
}
\arguments{
\item{...}{Any parameters for \code{\link[svglite]{svgstring}} function.}
}
\value{
No return value, the svg device is opened.
}
\description{
Open default svg device.
}
