% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throwBall.R
\name{throwBall}
\alias{throwBall}
\title{Throw a ball.}
\usage{
throwBall(distribution = "normal", param1 = 5, param2 = 1, posDF,
  step = Inf, distance = NULL)
}
\arguments{
\item{distribution}{string with distribution name, among:
'normal', 'uniform', 'poisson', 'bernoulli', 'binomial',
'geometric', 'degenerate', 'chisq', 'weibull'.}

\item{param1}{Numeric with first parameter from the distribution.}

\item{param2}{Numeric with second parameter from the distribution.}

\item{posDF}{Data.frame with ball positions.}

\item{step}{Numeric with step.
step = Inf preserves 'usual' animation, otherwise 'shiny' animation is used}

\item{distance}{Numeric with distance at which the ball is thrown.}
}
\value{
\code{posDF}with updated positions.
}
\description{
Throw a ball.
}
