% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineOutcome.R
\name{winnerNumber}
\alias{winnerNumber}
\title{Get the name of the winner.}
\usage{
winnerNumber(winner)
}
\arguments{
\item{winner}{String with name of winner}
}
\value{
String with name of winner with number.
}
\description{
Get the name of the winner.
}
